Enable CouchDB to manage OS process daemons.
This is a simple feature that allows useres to configure CouchDB
so that it maintains a given OS level process alive. If the process
dies for any reason, CouchDB will restart it. If the process restarts
too often, then CouchDB will mark it has halted and not attempt
to restart it. The default max restart rate is three times in the
last five seconds. These parameters are adjustable.

Commands that are started in this manner will have access to a simple
API over stdio to request configuration parameters or to add log
statements to CouchDB's logs.

To configure an OS process as a CouchDB os_daemon, create a section
in your local.ini like such:

    [os_daemons]
    daemon_name = /path/to/command -with args

This will make CouchDB bring up the command and attempt to keep it
alive. To request a configuration parameter, an os_daemon can write
a simple JSON message to stdout like such:

    ["get", "os_daemons"]\n

which would return:

    {"daemon_name": "/path/to/command -with args"}

Or:

    ["get", "os_daemons", "daemon_name"]\n

which would return:

    "/path/to/command -with args"

There's no restriction on what configuration variables are visible.
There's also no method for altering the configuration.

Logging commands look like:

    ["log", $(JSON_MESSAGE)]

Where $(JSON_MESSAGE) is arbitrary JSON data. These messages are
logged at the 'info' level.

When implementing a daemon process to be managed by CouchDB you
should remember to use a method like checking the parent process
id or if stdin has been closed. These flags can tell you if
your daemon process has been orphaned so you can exit cleanly.

Tests for this new feature are in:

tests/etap/170-*

Tests for the ability to control daemons from the config system.
Adding and removing config entries will dynamically bring the
daemons up and down.

tests/etap/171-*

These do some simple testsing of fetching values from the config
API's as well as provides a way for managed daemons to report
log messages that will be combined into CouchDB's log files.

tests/etap/172-*

Tests for various error conditions when a daemon is started. Errors
tests included files that are not executable, daemons that die
on boot, daemons that die shortly after boot, and daemons that die
but not often enough to be halted.