Implemented an HTTP proxy handler.
The second of two new features to replace the _externals protocols. This
allows users to configure CouchDB to proxy requests to an external HTTP
server. The external HTTP server is not required to be on the same host
running CouchDB.

The configuration looks like such:

[httpd_global_handlers]
_google = {couch_httpd_proxy, handle_proxy_req, <<"http://www.google.com">>}

You can then hit this proxy at the url:

http://127.0.0.1:5984/_google

If you add any path after the proxy name, or make a request with a query
string, those will be appended to the URL specified in the configuration.

Ie:

http://127.0.0.1:5984/_google/search?q=plankton

would translate to:

http://www.google.com/search?q=plankton

Request bodies are also proxied as expected.

It is also possible to use this to redirect internally to CouchDB. For
instance:

[httpd_global_handlers]
_foo = {couch_httpd_proxy, handle_proxy_req, <<"/other_db">>}

would basically alias '_foo' to the database 'other_db'. In general
you should prefer to use the other rewriting features of CouchDB as this
method creates a full HTTP request instead of just altering internal logic
of the server.

So far, the basic tests in test/etap/180-http-proxy.t check for various
parts of the HTTP request being proxied correctly. Still lacking is any
testing for chunked HTTP bodies.