BugzID: 11589 call file:close explicitly to keep VM responsive
Deleting a large file in the previous scheme has caused the Erlang VM
to become unresponsive (including making it drop out of the ring of
nodes).

The cause of this is currently unknown but Adam discovered that
explicitly calling file:close/1 on the file descriptor does not cause
this behavior.

Accordingly, I have typed in his suggested fix and tested it.

Conflicts:
	src/couch/src/couch_file.erl