Does Not Work
This was a passing attempt to remove file chunking from couch_file so
that data would be contiguous on disk. The basic idea is to use a
system similar to the current method but to only insert the zero
byte if a 1 byte happens to land on a header boundary.

If a 1 byte does hit that boundary when using normal writes, the
written binary is split into multiple chunks each time the a 1
byte hits a boundary. Each chunk is then written to disk as
we do currently with bytes inserted to prevent misidentified
headers.

The list of {Pos, Size} pairs for each chunk of the write is
then written at the end of the append with a zeroed out md5 to
indicate to readers to read the {Pos, Size} list and then
reconstruct the binary before sending it back to the reader.