Collapse dupes in views instead of crashing
Duplicate entries in a database's sequence tree can be introduced by the
bugs described in COUCHDB-968 and COUCHDB-1265.  The dupes will
propagate to any secondary indexes if they are sufficiently close enough
to each other in the seq tree.

This patch modifies the view compactor so that instead of crashing on
detection of dupes in the view index it collapses all the view keys into
a single entry for the document.  A subsequent indexing of the document
will remove the duped keys from the view and will replace them with the
correct output for the current revision of the doc.