Fix for the auth_cache etap
As it turns out, opening a doc by id is different than opening it using
a #doc_info record due to the inclusion of the full revision path. This
ended up breaking the auth_cache tests. This way includes the entire
revision path for all docs and not just first doc loads.