Improve the CLI JavaScript stack formating
This dresses up the output of the JavaScript tests by making sure the
output is sane. Most of the issue was that the tracebak prints function
arguments which can include the entire source of the test.

This also tries to print the source file and line number nicely so that
we can find right where errors have occurred. Its proven quite useful
while fixing JavaScript tests.