Fix race condition in oauth.js
On master this was relying on a terrible race condition to sneak a
DELETE of the temporary users db in after the config change. The auth
cache is now fast enough to reliably check the db exists before this
call which ends up in a fairly obvious 404 trying to create a users doc
in the auth db that was just deleted.

To fix we just delete the db before running on the modified server so
that its still cleaned out to the assumed empty state at the beginning
of the test.