[squash] 073-changes.t
This changed a lot cause of how Db's behave now in that you have to call
reopen to access records you've written. Ie, if you write a doc, and
then try and write a new revision without reopening you'll get a
conflict because the prepare_and_valdiate_* side is using old #btree{}
states that don't have the proper snapshot to extend the previous
revision.